+++ demo-ng/app/app.component.ts
@@ -1,8 +1,19 @@
-import { Component } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
+
+const firebase = require("nativescript-plugin-firebase");
 
 @Component({
-    selector: "ns-app",
-    templateUrl: "app.component.html",
+  selector: "ns-app",
+  templateUrl: "app.component.html",
 })
 
-export class AppComponent { }
+export class AppComponent implements OnInit {
+  ngOnInit(): void {
+    firebase.init(
+        {
+          persist: false
+        })
+        .then(() => console.log(">>>>> Firebase initialized"))
+        .catch(err => console.log(">>>>> Firebase init error: " + err));
+  }
+}