+++ demo-ng/app/item/items.component.ts
@@ -1,7 +1,11 @@
-import { Component, OnInit } from "@angular/core";
+import { Component, NgZone } from "@angular/core";
 import { firestore } from "nativescript-plugin-firebase";
+import { Observable } from "rxjs/Observable";
+import { City } from "../model/City";
+
 const firebase = require("nativescript-plugin-firebase/app");
 const firebaseWebApi = require("nativescript-plugin-firebase/app");
+
 // import { AngularFireModule } from 'angularfire2';
 
 @Component({
@@ -9,18 +13,17 @@ const firebaseWebApi = require("nativescript-plugin-firebase/app");
   moduleId: module.id,
   templateUrl: "./items.component.html",
 })
-export class ItemsComponent implements OnInit {
+export class ItemsComponent {
 
   private listenerUnsubscribe: () => void;
 
-  constructor() {
-    // AngularFireModule.initializeApp({});
-  }
+  public myCity$: Observable<City>;
+  public myCities$: Observable<Array<City>>;
+  private city: City;
+  private cities: Array<City> = [];
 
-  ngOnInit(): void {
-    firebase.initializeApp({
-      persist: false
-    }).then(() => console.log("Firebase initialized"));
+  constructor(private zone: NgZone) {
+    // AngularFireModule.initializeApp({});
   }
 
   public loginAnonymously(): void {
@@ -122,6 +125,9 @@ export class ItemsComponent implements OnInit {
         .then((querySnapshot: firestore.QuerySnapshot) => {
           querySnapshot.forEach(doc => {
             console.log(`${doc.id} => ${JSON.stringify(doc.data())}`);
+            // since there's a reference stored here, we can use that to retrieve its data
+            const docRef: firestore.DocumentReference = doc.data().ref2sf;
+            docRef.get().then(res => console.log("docref.get: " + JSON.stringify(res.data())));
           });
         })
         .catch(err => console.log("Get failed, error" + err));
@@ -147,7 +153,7 @@ export class ItemsComponent implements OnInit {
             .doc("SF")
             .collection("streets")
             .doc("QZNrg22tkN8W71YC3qCb"); // id of 'main st.'
-            // .doc("doesntexist");
+    // .doc("doesntexist");
 
     mainStreetInSFDocRef.get().then((doc: firestore.DocumentSnapshot) => {
       if (doc.exists) {
@@ -157,6 +163,32 @@ export class ItemsComponent implements OnInit {
       }
     }).catch(function (error) {
       console.log("Error getting document:", error);
+    });
+  }
+
+
+  firestoreDocumentObservable(): void {
+    this.myCity$ = Observable.create(subscriber => {
+      const docRef: firestore.DocumentReference = firebase.firestore().collection("cities").doc("SF");
+      docRef.onSnapshot((doc: firestore.DocumentSnapshot) => {
+        this.zone.run(() => {
+          this.city = <City>doc.data();
+          subscriber.next(this.city);
+        });
+      });
+    });
+  }
+
+  firestoreCollectionObservable(): void {
+    this.myCities$ = Observable.create(subscriber => {
+      const colRef: firestore.CollectionReference = firebase.firestore().collection("cities");
+      colRef.onSnapshot((snapshot: firestore.QuerySnapshot) => {
+        this.zone.run(() => {
+          this.cities = [];
+          snapshot.forEach(docSnap => this.cities.push(<City>docSnap.data()));
+          subscriber.next(this.cities);
+        });
+      });
     });
   }